requirements.txt
streamlit==1.23.1
streamlit_chat==0.0.2.2
langchain==0.0.191
openai==0.27.6
faiss-cpu==1.7.4
tiktoken==0.4.0
pandasai==0.5.5
pandas_profiling==3.6.3
treamlit-pandas-profiling==0.1.3
